requirements.txt
torch>=1.9.0
torchaudio>=0.9.0
numpy==2.0.0
scipy>=1.7.0
librosa==0.10.1
numba<0.58.0
tensorboard>=2.6.0
matplotlib>=3.4.0
pandas>=1.3.0
scikit-learn>=1.0.0
soundfile>=0.10.0
tqdm>=4.62.0
PyYAML>=6.0